refactor(post): useGetPosts 훅을 post-list로 이동 및 관련 코드 정리
