+++ src/entities/post/api/index.ts
@@ -3,5 +3,3 @@ export { getPosts, searchPosts, getPostsByTag } from "./services"
 
 // Query Keys
 export { postQueryKeys } from "./query-keys"
-
-export { useGetPosts } from "./hooks/use-get-posts"
